convert visplane_t.top/bottom to unsigned short

+++ src/r_defs.h
@@ -437,12 +437,11 @@ typedef struct visplane
   int picnum, lightlevel, minx, maxx;
   fixed_t height;
   fixed_t xoffs, yoffs;         // killough 2/28/98: Support scrolling flats
-  // [FG] 32-bit integer math
-  unsigned int pad1;          // leave pads for [minx-1]/[maxx+1]
-  unsigned int top[MAX_SCREENWIDTH];
-  unsigned int pad2, pad3;    // killough 2/8/98, 4/25/98
-  unsigned int bottom[MAX_SCREENWIDTH];
-  unsigned int pad4;
+  unsigned short pad1;          // leave pads for [minx-1]/[maxx+1]
+  unsigned short top[MAX_SCREENWIDTH];
+  unsigned short pad2, pad3;    // killough 2/8/98, 4/25/98
+  unsigned short bottom[MAX_SCREENWIDTH];
+  unsigned short pad4;
 } visplane_t;
 
 #endif
